requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
dacite==1.8.1
haggis==0.13.0
idna==3.7
lxml==5.2.2
markdownify==0.12.1
marko==2.1.2
numpy==2.0.0
python-docx==1.1.2
requests==2.32.3
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2